Use ULONG_MAX in MAX_BLOCK_SIZE

+++ includes/malloc.h
@@ -73,7 +73,7 @@ typedef struct		s_zone
 
 # define LARGE_THRESHOLD (SMALL_THRESHOLD + 1)
 
-# define MAX_BLOCK_SIZE (UINT_MAX - LARGE_ZONE_HEADER_SIZE)
+# define MAX_BLOCK_SIZE (ULONG_MAX - LARGE_ZONE_HEADER_SIZE)
 
 typedef struct		s_zones
 {
